Refactor(Alerta): Se movió de ubicación
- Se movieron los archivos Alerta.js y alerta.css dentro de la carpeta Alerta
